rebase to current state of main, move meta get tests into integration_tests file
